feat(k0s): Add support for persistence storage via CSI
In default configurations of k0s, persistence storage is disabled. k0s comes
with an implementation for container storage interfaces and OpenESB is set as
the default for that purpose.

With this commit, we can now enable or disable persistent storages and switch
to different CSI implementations (e.g. Minio) when wanted.

feat(k0s): Add k0sctl

feat(k0s): Fallback to stable k0s_version when not defined